requirements.txt
# Core Framework
streamlit>=1.30.0

# Database
sqlalchemy>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
plotly>=5.17.0
pyvis>=0.3.2
seaborn>=0.12.0
matplotlib>=3.7.0

# Network Analysis
networkx>=3.0

# Utilities
python-dotenv>=1.0.0
faker>=20.0.0
pydantic>=2.0.0
